add 'new mode' which simplifies margins
overview
--------

In the existing code, there are three parameters (normal/min/aligned
leader), which control gap sizes at the start and end of the document.
It's not really clear to me what the intent behind these are, or how to
set them up to make things align nicely. This is an attempt to tidy this
up by just specifying the parameters of the printer, rather than what
gaps to add where.

There are two main parameters which we need to make things work
properly:

- HCD: the distance from the head to the cutter
- margin: the amount of margin to add if the cut might be close to the
  printed region

At the start of the print, there will always be a HCD-length gap (from
cutting the previous label), but at the end of the print the amount of
gap is variable -- it will be 0 if we advance by HCD before cutting.

In center mode, there's no need to add an extra gap at the start because
we already have the HCD gap, so overall the procedure is

- send document
- skip HCD + HCD
- cut

In right (end) aligned mode, we instead add the margin to avoid cutting
the end off:

- send document
- skip HCD + margin
- cut

Left (start) aligned mode works the same, but the document is reversed
to put the gap at the other end:

- send document reversed
- skip HCD + margin
- cut

limitations
-----------

Continuous and auto paper modes will not do anything if this is enabled;
perhaps we should fall back to the old behaviour if these are set? The
main problem here is that I don't really understand what these modes do.

'Chain Marks' don't seem to work in new mode, but they didn't seem to
work properly in old mode either. It shouldn't be a problem,

The minimum label length is not implemented. I don't think this is that
useful, but if it's to be implemented it should at least make the gaps
symmetrical in center mode; this probably means that the whole page
should always be buffered.

It would be good to add an 'exact mode' in which no gaps are added (but
content might be cut off), to make exact-length labels more predictable.